+++ src/lib/components/forms/form/form-components/text-input/text-input.component.ts
@@ -1,6 +1,9 @@
+import { Component, Inject, Input } from '@angular/core';
 import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
-import { Component, Input } from '@angular/core';
 import { TextInputInterface } from './text-input.interface';
+import { TerraPlacementEnum } from '../../../../../helpers/enums/terra-placement.enum';
+import { noop } from 'rxjs';
+import { L10N_LOCALE, L10nLocale } from 'angular-l10n';
 
 @Component({
     selector: 'tc-text-input',
@@ -14,41 +17,87 @@ import { TextInputInterface } from './text-input.interface';
     ]
 })
 export class TextInputComponent implements ControlValueAccessor, TextInputInterface {
+    /**
+     * @description If true, the type of input will be 'password'.
+     * @default false
+     */
     @Input()
-    public isDisabled: boolean;
+    public isPassword: boolean = false;
 
+    /**
+     * @description If true, the input will check if the input is a valid iban.
+     * @default false
+     */
     @Input()
-    public isIban: boolean;
+    public isIban: boolean = false;
 
+    /**
+     * @description If true, the value cannot be changed.
+     * @default false
+     */
     @Input()
-    public isPassword: boolean;
+    public isReadonly: boolean = false;
 
+    /**
+     * @description If true, the button will be disabled.
+     * @Default false.
+     * */
     @Input()
-    public isReadonly: boolean;
+    public isDisabled: boolean = false;
 
+    /**
+     * @description If true, a * indicates that the value is required.
+     *  @Default false.
+     *  */
     @Input()
-    public isRequired: boolean;
+    public isRequired: boolean = false;
 
+    /** @description Set the maximum number value allowed. */
     @Input()
     public maxLength: number;
 
+    /** @description Set the minimum number value allowed. */
     @Input()
     public minLength: number;
 
+    /** @description Set the label. */
     @Input()
-    public name: string;
+    public name: string = '';
 
+    /** @description Set the tooltip. */
     @Input()
-    public tooltipPlacement: string;
+    public tooltipText: string = '';
 
+    /**
+     * @description Set the tooltip placement (bottom, top, left, right).
+     * @default top
+     * */
     @Input()
-    public tooltipText: string;
+    public tooltipPlacement: TerraPlacementEnum = TerraPlacementEnum.TOP;
 
-    constructor() {}
+    /** @description The internal data model */
+    public value: string;
 
-    public registerOnChange(fn: any): void {}
+    /** Stores a callback function which is executed whenever the input was blurred. */
+    public _onTouchedCallback: () => void = noop;
 
-    public registerOnTouched(fn: any): void {}
+    /** Stores a callback function which is executed whenever the value of the input changes. */
+    public _onChangeCallback: (_: string) => void = noop;
 
-    public writeValue(obj: any): void {}
+    constructor(@Inject(L10N_LOCALE) public _locale: L10nLocale) {}
+
+    /** @description Registers a callback function that is called when the control's value changes in the UI.*/
+    public registerOnChange(fn: (_: string) => void): void {
+        this._onChangeCallback = fn;
+    }
+
+    /** @description Registers a callback function that is called by the forms API on initialization to update the form model on blur. */
+    public registerOnTouched(fn: () => void): void {
+        this._onTouchedCallback = fn;
+    }
+
+    /** @description Writes a new value to the element.*/
+    public writeValue(value: string): void {
+        this.value = value;
+    }
 }